변경 내용 저장

+++ backend/src/main/java/com/tournet/tournetERP/common/dto/ComCodeDTO.java
@@ -18,9 +18,9 @@ public class ComCodeDTO {
 
     private String codeEn;
 
-    private int codeLvl;
+    private long codeLvl;
 
-    private int codeOrd;
+    private long codeOrd;
 
     private long uprCodeUuid;
